Rename services module to service
